# apiserver/src/main/resources/bob/api.yaml
openapi: "3.0.0"
info:
  title: Bob the Builder
  version: "0.1.0"
  description: The modular, extensible CI/CD platform
paths:
...
  /can-we-build-it:
    get:
      operationId: HealthCheck
      summary: Runs health checks for Bob

      responses:
        "200":
          description: Successful health check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleResponse"
              examples:
                healthy:
                  value:
                    message: "Yes we can! 🔨 🔨"

        default:
          description: Failed health check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleResponse"
              examples:
                unhealthy:
                  value:
                    message: "Health check failed: Postgres not healthy"
...
components:
  schemas:
    SimpleResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...